Address organizational and namespace comments [skip ci]

+++ tiledb/stdx/__ranges/test/unit_chunk_view.cc
@@ -31,7 +31,7 @@
  */
 
 #include <catch2/catch_all.hpp>
-#include "tiledb/common/util/chunk_view.h"
+#include "tiledb/stdx/__ranges/chunk_view.h"
 
 TEST_CASE("chunk_view: null test", "[chunk_view][null test]") {
   REQUIRE(true);
@@ -359,8 +359,8 @@ TEST_CASE("chunk_view: Iterators - 1", "[chunk_view]") {
 
 TEST_CASE("chunk_view: Iterators", "[chunk_view]") {
   for (auto&& v : {v10, v11, v12, v13, v14, v15, v16}) {
-    for (long i = 1; i <= v.size(); ++i) {
-      auto a = _cpo::chunk(v, i);
+    for (size_t i = 1; i <= v.size(); ++i) {
+      auto a = _cpo::chunk(v, (long) i);
       auto b = a.begin();
       CHECK(b->size() == (unsigned)i);
     }
